[wr_streamers] bugfix: 0xCAFE as last word caused streamers to hang
Fixed by adding missing "else statement" in CRC_WORD state. In this
statement the dvalid is set LOW and no ('X') data is set. Without
this, the input data remainded 0xCAFE and the input dvalid remainded
HIGH, thus the escape_inserter was forcing dreq=LOW in order to
stop the input data for one cycle and insert special character.